+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.5.0 (IN PROGRESS)
 
 * Fix `craftLayerUrl()` loading state (FOLIO-1547)
+* Replace `formatDateTime()` with `<FormattedTime>` (STCOR-267)
 
 ## [1.4.0](https://github.com/folio-org/ui-inventory/tree/v1.4.0) (2018-10-05)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.3.0...v1.4.0)